Re-fix bug
Don't fix bugs late at night! I spotted the wrong thing. In actuality, I'd introduced a bug whereby the field headings would only be the last set of keys larger than the first set of keys. Generally this was fine, but unfortunately there was an edge case when the first set of keys was very small - because the participant had no connections - and the last set of keys was smaller than the largest.

I've done away with the ugly for loop and just settled on a nice max with key