Move weightings to process module
This change causes the Match object to demand a weightings object. It moves the default values up to the process module, making it clearer what they are, and opening the way for alterations to these weightings in future releases.

This should not be a breaking change. The mechanics of how matching and processing occur are the same.