Remove old approach to scoring
This significant change removes the old scoring methods from the `Match` object. Instead, `Rule` objects must now be passed to the `Match` object.

I've written a little bit of code that replicates the pre-existing rules. This should allow clients to continue using the software as it was before this breaking change, as long as they don't call the `Match` object directly.